car_frontend-master/backend/requirements.txt
bidict==0.21.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
Cython==0.29.23
Flask==2.0.1
Flask-Cors==3.0.10
Flask-SocketIO==5.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pybind11==2.6.2
numpy==1.20.3
python-engineio==4.2.0
python-socketio==5.3.0
pyzmq==22.1.0
requests==2.25.1
scipy==1.6.3
six==1.16.0
urllib3==1.26.5
Werkzeug==2.0.1
